requirements.txt
schwab-py>=1.4.0
anthropic>=0.40.0
python-dotenv>=1.0.0
yfinance>=0.2.0
